Update based on review comments
