+++ Utils/hdinsight-node-common/src/com/microsoft/azure/hdinsight/sdk/rest/azure/projectarcadia/models/ConnectivityEndpoints.java
@@ -21,13 +21,15 @@
  *
  */
 
-package com.microsoft.azure.hdinsight.sdk.rest.arcadia.models;
+package com.microsoft.azure.hdinsight.sdk.rest.azure.projectarcadia.models;
 
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 /**
  * Workspace connectivity endpoints.
  */
+@JsonIgnoreProperties(ignoreUnknown = true)
 public class ConnectivityEndpoints {
     /**
      * Web endpoint.
