+++ Utils/hdinsight-node-common/src/com/microsoft/azure/hdinsight/sdk/rest/azure/projectarcadia/models/ErrorResponse.java
@@ -21,14 +21,17 @@
  *
  */
 
-package com.microsoft.azure.hdinsight.sdk.rest.arcadia.models;
+package com.microsoft.azure.hdinsight.sdk.rest.azure.projectarcadia.models;
 
 import java.util.List;
+
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 /**
  * The error response.
  */
+@JsonIgnoreProperties(ignoreUnknown = true)
 public class ErrorResponse {
     /**
      * The error code.
